requirements.txt
requests
pandas
qrcode
pillow
matplotlib
